+++ renderer/src/components/main-panel/chat/core/task-loop.ts
@@ -3,10 +3,10 @@ import { ref, type Ref } from "vue";
 import { type ToolCall, type ChatStorage, getToolSchema, MessageState } from "../chat-box/chat";
 import { useMessageBridge, MessageBridge, createMessageBridge } from "@/api/message-bridge";
 import type { OpenAI } from 'openai';
-import { llmManager, llms } from "@/views/setting/llm";
+import { llmManager, llms, type BasicLlmDescription } from "@/views/setting/llm";
 import { pinkLog, redLog } from "@/views/setting/util";
 import { ElMessage } from "element-plus";
-import { handleToolCalls, type ToolCallResult } from "./handle-tool-calls";
+import { getToolCallIndexAdapter, handleToolCalls, type IToolCallIndex, type ToolCallResult } from "./handle-tool-calls";
 import { getPlatform } from "@/api/platform";
 import { getSystemPrompt } from "../chat-box/options/system-prompt";
 import { mcpSetting } from "@/hook/mcp";
@@ -45,7 +45,7 @@ export class TaskLoop {
     private onToolCalled: (toolCallResult: ToolCallResult) => ToolCallResult = toolCallResult => toolCallResult;
     private onEpoch: () => void = () => {};
     private completionUsage: ChatCompletionChunk['usage'] | undefined;
-    private llmConfig: any;
+    private llmConfig?: BasicLlmDescription;
 
     constructor(
         private readonly taskOptions: TaskLoopOptions = { maxEpochs: 20, maxJsonParseRetry: 3, adapter: undefined },
@@ -76,7 +76,7 @@ export class TaskLoop {
         }
     }
 
-    private handleChunkDeltaToolCalls(chunk: ChatCompletionChunk) {
+    private handleChunkDeltaToolCalls(chunk: ChatCompletionChunk, toolcallIndexAdapter: (toolCall: ToolCall) => IToolCallIndex) {
         const toolCall = chunk.choices[0]?.delta?.tool_calls?.[0];
         
         if (toolCall) {
@@ -84,7 +84,8 @@ export class TaskLoop {
                 console.warn('tool_call.index is undefined or null');
             }
 
-            const index = toolCall.index || 0;
+
+            const index = toolcallIndexAdapter(toolCall);
             const currentCall = this.streamingToolCalls.value[index];
 
             if (currentCall === undefined) {
@@ -105,10 +106,10 @@ export class TaskLoop {
                         currentCall.id = toolCall.id;
                     }
                     if (toolCall.function?.name) {
-                        currentCall.function.name = toolCall.function.name;
+                        currentCall.function!.name = toolCall.function.name;
                     }
                     if (toolCall.function?.arguments) {
-                        currentCall.function.arguments += toolCall.function.arguments;                        
+                        currentCall.function!.arguments += toolCall.function.arguments;                        
                     }
                 }
             }
@@ -123,7 +124,7 @@ export class TaskLoop {
         }
     }
 
-    private doConversation(chatData: ChatCompletionCreateParamsBase) {
+    private doConversation(chatData: ChatCompletionCreateParamsBase, toolcallIndexAdapter: (toolCall: ToolCall) => IToolCallIndex) {
 
         return new Promise<IDoConversationResult>((resolve, reject) => {
             const chunkHandler = this.bridge.addCommandListener('llm/chat/completions/chunk', data => {
@@ -134,7 +135,7 @@ export class TaskLoop {
 
                 // 处理增量的 content 和 tool_calls
                 this.handleChunkDeltaContent(chunk);
-                this.handleChunkDeltaToolCalls(chunk);
+                this.handleChunkDeltaToolCalls(chunk, toolcallIndexAdapter);
                 this.handleChunkUsage(chunk);
                 
                 this.onChunk(chunk);
@@ -352,9 +353,11 @@ export class TaskLoop {
             }
 
             this.currentChatId = chatData.id!;
+            const llm = this.getLlmConfig();
+            const toolcallIndexAdapter = getToolCallIndexAdapter(llm);
 
             // 发送请求
-            const doConverationResult = await this.doConversation(chatData);
+            const doConverationResult = await this.doConversation(chatData, toolcallIndexAdapter);
 
             console.log('[doConverationResult] Response');
             console.log(doConverationResult);
@@ -405,7 +408,7 @@ export class TaskLoop {
                     } else if (toolCallResult.state === MessageState.Success) {
                         tabStorage.messages.push({
                             role: 'tool',
-                            index: toolCall.index || 0,
+                            index: toolcallIndexAdapter(toolCall),
                             tool_call_id: toolCall.id || '',
                             content: toolCallResult.content,
                             extraInfo: {
@@ -419,8 +422,8 @@ export class TaskLoop {
     
                         tabStorage.messages.push({
                             role: 'tool',
-                            index: toolCall.index || 0,
-                            tool_call_id: toolCall.id || toolCall.function.name,
+                            index: toolcallIndexAdapter(toolCall),
+                            tool_call_id: toolCall.id || toolCall.function!.name,
                             content: toolCallResult.content,
                             extraInfo: {
                                 created: Date.now(),