+++ service/src/llm/llm.controller.ts
@@ -1,7 +1,7 @@
-import { Controller, RequestClientType } from "../common";
+import { OpenAI } from "openai";
+import { Controller } from "../common";
 import { RequestData } from "../common/index.dto";
 import { PostMessageble } from "../hook/adapter";
-import { getClient } from "../mcp/connect.service";
 import { abortMessageService, streamingChatCompletion } from "./llm.service";
 
 export class LlmController {
@@ -10,7 +10,7 @@ export class LlmController {
     async chatCompletion(data: RequestData, webview: PostMessageble) {
 
         try {
-            await streamingChatCompletion(data, webview);            
+            await streamingChatCompletion(data, webview);
         } catch (error) {
             console.log('error' + error);
             
@@ -34,4 +34,20 @@ export class LlmController {
         return abortMessageService(data, webview);
     }
 
+
+    @Controller('llm/models')
+    async getModels(data: RequestData, webview: PostMessageble) {
+        const {
+            baseURL,
+            apiKey,
+        } = data;        
+        
+        const client = new OpenAI({ apiKey, baseURL });
+        const models = await client.models.list();
+
+        return {
+            code: 200,
+            msg: models.data
+        }
+    }
 }